requirements.txt
fastapi
uvicorn[standard]
pandas
python-multipart
python-dotenv
requests
passlib[bcrypt]
python-jose[cryptography]